Append AccessGrants to SocialAgentRegistrations
