+++ pod-provider/backend/services/sai/agent-registry/app-registrations.js
@@ -1,3 +1,7 @@
+const path = require('path');
+const urlJoin = require('url-join');
+const { MoleculerError } = require('moleculer').Errors;
+const { ACTIVITY_TYPES } = require('@semapps/activitypub');
 const { ControlledContainerMixin } = require('@semapps/ldp');
 const { MIME_TYPES } = require('@semapps/mime-types');
 const AgentRegistrationsMixin = require('../../../mixins/agent-registrations');
@@ -15,6 +19,26 @@ module.exports = {
     excludeFromMirror: true,
     activateTombstones: false,
     typeIndex: 'private'
+  },
+  dependencies: ['api', 'ldp'],
+  async started() {
+    const basePath = await this.broker.call('ldp.getBasePath');
+    await this.broker.call('api.addRoute', {
+      route: {
+        name: 'auth-agent-app-registration',
+        path: path.join(basePath, '/.auth-agent/app-registrations'),
+        authorization: true,
+        authentication: false,
+        bodyParsers: {
+          json: true
+        },
+        aliases: {
+          'POST /register': 'app-registrations.register',
+          'POST /upgrade': 'app-registrations.upgrade',
+          'POST /remove': 'app-registrations.remove'
+        }
+      }
+    });
   },
   actions: {
     async createOrUpdate(ctx) {
@@ -73,6 +97,143 @@ module.exports = {
         );
 
         return appRegistrationUri;
+      }
+    },
+    async register(ctx) {
+      let { appUri, acceptedAccessNeeds, acceptedSpecialRights, acceptAllRequirements = false } = ctx.params;
+
+      const webId = ctx.meta.webId;
+      const account = await ctx.call('auth.account.findByWebId', { webId });
+      ctx.meta.dataset = account.username;
+
+      // Force to get through network
+      const app = await ctx.call('ldp.remote.getNetwork', { resourceUri: appUri });
+
+      const appRegistration = await ctx.call('app-registrations.getForAgent', { agentUri: appUri, podOwner: webId });
+
+      if (appRegistration) {
+        throw new MoleculerError(
+          `User already has an application registration. Upgrade or uninstall the app first.`,
+          400,
+          'BAD REQUEST'
+        );
+      }
+
+      if (acceptAllRequirements) {
+        if (acceptedAccessNeeds || acceptedSpecialRights) {
+          throw new Error(
+            `If acceptAllRequirements is true, you should not pass acceptedAccessNeeds or acceptedSpecialRights`
+          );
+        }
+
+        const requirements = await ctx.call('applications.getRequirements', { appUri });
+        acceptedAccessNeeds = requirements.accessNeeds;
+        acceptedSpecialRights = requirements.specialRights;
+      }
+
+      const appRegistrationUri = await ctx.call('app-registrations.createOrUpdate', {
+        appUri,
+        podOwner: webId,
+        acceptedAccessNeeds,
+        acceptedSpecialRights
+      });
+
+      if (this.broker.cacher) {
+        // Invalidate all rights of the application on the Pod as they may now be completely different
+        await ctx.call('webacl.cache.invalidateAllUserRightsOnPod', { webId: appUri, podOwner: webId });
+      }
+
+      // If the app is an ActivityPub actor, send notification
+      if (app.inbox) {
+        await ctx.call('activitypub.outbox.post', {
+          collectionUri: urlJoin(webId, 'outbox'),
+          '@type': ACTIVITY_TYPES.CREATE,
+          object: appRegistrationUri,
+          to: appUri
+        });
+      }
+
+      return appRegistrationUri;
+    },
+    async upgrade(ctx) {
+      let { appUri, acceptedAccessNeeds, acceptedSpecialRights, acceptAllRequirements = false } = ctx.params;
+
+      const webId = ctx.meta.webId;
+      const account = await ctx.call('auth.account.findByWebId', { webId });
+      ctx.meta.dataset = account.username;
+
+      // Force to get through network
+      const app = await ctx.call('ldp.remote.getNetwork', { resourceUri: appUri });
+
+      if (acceptAllRequirements) {
+        if (acceptedAccessNeeds || acceptedSpecialRights) {
+          throw new Error(
+            `If acceptAllRequirements is true, you should not pass acceptedAccessNeeds or acceptedSpecialRights`
+          );
+        }
+
+        const requirements = await ctx.call('applications.getRequirements', { appUri });
+        acceptedAccessNeeds = requirements.accessNeeds;
+        acceptedSpecialRights = requirements.specialRights;
+      }
+
+      const appRegistrationUri = await ctx.call('app-registrations.createOrUpdate', {
+        appUri,
+        podOwner: webId,
+        acceptedAccessNeeds,
+        acceptedSpecialRights
+      });
+
+      if (this.broker.cacher) {
+        // Invalidate all rights of the application on the Pod as they may now be completely different
+        await ctx.call('webacl.cache.invalidateAllUserRightsOnPod', { webId: appUri, podOwner: webId });
+      }
+
+      // If the app is an ActivityPub actor, send notification
+      if (app.inbox) {
+        await ctx.call('activitypub.outbox.post', {
+          collectionUri: urlJoin(webId, 'outbox'),
+          '@type': ACTIVITY_TYPES.UPDATE,
+          object: appRegistrationUri,
+          to: appUri
+        });
+      }
+
+      return appRegistrationUri;
+    },
+    async remove(ctx) {
+      const { appUri } = ctx.params;
+
+      const webId = ctx.meta.webId;
+      const account = await ctx.call('auth.account.findByWebId', { webId });
+      ctx.meta.dataset = account.username;
+
+      const app = await ctx.call('applications.get', { appUri, webId });
+      const appRegistration = await ctx.call('app-registrations.getForAgent', { agentUri: appUri, podOwner: webId });
+
+      if (appRegistration) {
+        // Immediately delete existing webhooks channels to avoid permissions errors later
+        await ctx.call('solid-notifications.provider.webhook.deleteAppChannels', { appUri, webId });
+
+        await ctx.call('app-registrations.delete', {
+          resourceUri: appRegistration.id,
+          webId
+        });
+
+        // If the app is an ActivityPub actor, send notification
+        if (app.inbox) {
+          await ctx.call('activitypub.outbox.post', {
+            collectionUri: urlJoin(webId, 'outbox'),
+            type: ACTIVITY_TYPES.DELETE,
+            object: appRegistration.id || appRegistration['@id'],
+            to: appUri
+          });
+        }
+
+        if (this.broker.cacher) {
+          // Invalidate all rights of the application on the Pod as they may now be completely different
+          await ctx.call('webacl.cache.invalidateAllUserRightsOnPod', { webId: appUri, podOwner: webId });
+        }
       }
     }
   },
