+++ pod-provider/backend/services/sai/auth-agent.js
@@ -1,9 +1,6 @@
-const path = require('path');
 const urlJoin = require('url-join');
 const { triple, namedNode } = require('@rdfjs/data-model');
-const { MoleculerError } = require('moleculer').Errors;
 const { SingleResourceContainerMixin, getWebIdFromUri } = require('@semapps/ldp');
-const { ACTIVITY_TYPES } = require('@semapps/activitypub');
 const CONFIG = require('../../config/config');
 
 module.exports = {
@@ -22,26 +19,6 @@ module.exports = {
     }
   },
   dependencies: ['api', 'ldp'],
-  async started() {
-    const basePath = await this.broker.call('ldp.getBasePath');
-    await this.broker.call('api.addRoute', {
-      route: {
-        name: 'auth-agent',
-        path: path.join(basePath, '/.auth-agent'),
-        authorization: true,
-        authentication: false,
-        bodyParsers: {
-          json: true
-        },
-        aliases: {
-          'POST /register-app': 'auth-agent.registerApp',
-          'POST /upgrade-app': 'auth-agent.upgradeApp',
-          'POST /remove-app': 'auth-agent.removeApp',
-          'POST /share-resource': 'auth-agent.shareResource'
-        }
-      }
-    });
-  },
   actions: {
     // Action from the ControlledContainerMixin, called when we do GET or HEAD requests on resources
     async getHeaderLinks(ctx) {
@@ -67,167 +44,6 @@ module.exports = {
             rel: 'http://www.w3.org/ns/solid/interop#registeredAgent'
           }
         ];
-      }
-    },
-    async registerApp(ctx) {
-      let { appUri, acceptedAccessNeeds, acceptedSpecialRights, acceptAllRequirements = false } = ctx.params;
-
-      const webId = ctx.meta.webId;
-      const account = await ctx.call('auth.account.findByWebId', { webId });
-      ctx.meta.dataset = account.username;
-
-      // Force to get through network
-      const app = await ctx.call('ldp.remote.getNetwork', { resourceUri: appUri });
-
-      const appRegistration = await ctx.call('app-registrations.getForAgent', { agentUri: appUri, podOwner: webId });
-
-      if (appRegistration) {
-        throw new MoleculerError(
-          `User already has an application registration. Upgrade or uninstall the app first.`,
-          400,
-          'BAD REQUEST'
-        );
-      }
-
-      if (acceptAllRequirements) {
-        if (acceptedAccessNeeds || acceptedSpecialRights) {
-          throw new Error(
-            `If acceptAllRequirements is true, you should not pass acceptedAccessNeeds or acceptedSpecialRights`
-          );
-        }
-
-        const requirements = await ctx.call('applications.getRequirements', { appUri });
-        acceptedAccessNeeds = requirements.accessNeeds;
-        acceptedSpecialRights = requirements.specialRights;
-      }
-
-      const appRegistrationUri = await ctx.call('app-registrations.createOrUpdate', {
-        appUri,
-        podOwner: webId,
-        acceptedAccessNeeds,
-        acceptedSpecialRights
-      });
-
-      if (this.broker.cacher) {
-        // Invalidate all rights of the application on the Pod as they may now be completely different
-        await ctx.call('webacl.cache.invalidateAllUserRightsOnPod', { webId: appUri, podOwner: webId });
-      }
-
-      // If the app is an ActivityPub actor, send notification
-      if (app.inbox) {
-        await ctx.call('activitypub.outbox.post', {
-          collectionUri: urlJoin(webId, 'outbox'),
-          '@type': ACTIVITY_TYPES.CREATE,
-          object: appRegistrationUri,
-          to: appUri
-        });
-      }
-
-      return appRegistrationUri;
-    },
-    async upgradeApp(ctx) {
-      let { appUri, acceptedAccessNeeds, acceptedSpecialRights, acceptAllRequirements = false } = ctx.params;
-
-      const webId = ctx.meta.webId;
-      const account = await ctx.call('auth.account.findByWebId', { webId });
-      ctx.meta.dataset = account.username;
-
-      // Force to get through network
-      const app = await ctx.call('ldp.remote.getNetwork', { resourceUri: appUri });
-
-      if (acceptAllRequirements) {
-        if (acceptedAccessNeeds || acceptedSpecialRights) {
-          throw new Error(
-            `If acceptAllRequirements is true, you should not pass acceptedAccessNeeds or acceptedSpecialRights`
-          );
-        }
-
-        const requirements = await ctx.call('applications.getRequirements', { appUri });
-        acceptedAccessNeeds = requirements.accessNeeds;
-        acceptedSpecialRights = requirements.specialRights;
-      }
-
-      const appRegistrationUri = await ctx.call('app-registrations.createOrUpdate', {
-        appUri,
-        podOwner: webId,
-        acceptedAccessNeeds,
-        acceptedSpecialRights
-      });
-
-      if (this.broker.cacher) {
-        // Invalidate all rights of the application on the Pod as they may now be completely different
-        await ctx.call('webacl.cache.invalidateAllUserRightsOnPod', { webId: appUri, podOwner: webId });
-      }
-
-      // If the app is an ActivityPub actor, send notification
-      if (app.inbox) {
-        await ctx.call('activitypub.outbox.post', {
-          collectionUri: urlJoin(webId, 'outbox'),
-          '@type': ACTIVITY_TYPES.UPDATE,
-          object: appRegistrationUri,
-          to: appUri
-        });
-      }
-
-      return appRegistrationUri;
-    },
-    async removeApp(ctx) {
-      const { appUri } = ctx.params;
-
-      const webId = ctx.meta.webId;
-      const account = await ctx.call('auth.account.findByWebId', { webId });
-      ctx.meta.dataset = account.username;
-
-      const app = await ctx.call('applications.get', { appUri, webId });
-      const appRegistration = await ctx.call('app-registrations.getForAgent', { agentUri: appUri, podOwner: webId });
-
-      if (appRegistration) {
-        // Immediately delete existing webhooks channels to avoid permissions errors later
-        await ctx.call('solid-notifications.provider.webhook.deleteAppChannels', { appUri, webId });
-
-        await ctx.call('app-registrations.delete', {
-          resourceUri: appRegistration.id,
-          webId
-        });
-
-        // If the app is an ActivityPub actor, send notification
-        if (app.inbox) {
-          await ctx.call('activitypub.outbox.post', {
-            collectionUri: urlJoin(webId, 'outbox'),
-            type: ACTIVITY_TYPES.DELETE,
-            object: appRegistration.id || appRegistration['@id'],
-            to: appUri
-          });
-        }
-
-        if (this.broker.cacher) {
-          // Invalidate all rights of the application on the Pod as they may now be completely different
-          await ctx.call('webacl.cache.invalidateAllUserRightsOnPod', { webId: appUri, podOwner: webId });
-        }
-      }
-    },
-    async shareResource(ctx) {
-      const { resourceUri, authorizations } = ctx.params;
-
-      const podOwner = ctx.meta.webId;
-      const account = await ctx.call('auth.account.findByWebId', { webId: podOwner });
-      ctx.meta.dataset = account.username;
-
-      for ({ grantee, accessModes } of authorizations) {
-        if (accessModes.length > 0) {
-          await ctx.call('access-authorizations.generateForSingleResource', {
-            resourceUri,
-            podOwner,
-            grantee,
-            accessModes
-          });
-        } else {
-          await ctx.call('access-authorizations.deleteForSingleResource', {
-            resourceUri,
-            podOwner,
-            grantee
-          });
-        }
       }
     }
   },
