+++ pod-provider/backend/services/sai/auth-registry/data-grants.js
@@ -1,8 +1,9 @@
 const { ControlledContainerMixin } = require('@semapps/ldp');
+const ImmutableContainerMixin = require('../../../mixins/immutable-container-mixin');
 
 module.exports = {
   name: 'data-grants',
-  mixins: [ControlledContainerMixin],
+  mixins: [ControlledContainerMixin, ImmutableContainerMixin],
   settings: {
     acceptedTypes: ['interop:DataGrant'],
     newResourcesPermissions: {
@@ -16,12 +17,6 @@ module.exports = {
   },
   dependencies: ['ldp', 'ldp.registry'],
   actions: {
-    put() {
-      throw new Error(`The resources of type interop:DataGrant are immutable`);
-    },
-    patch() {
-      throw new Error(`The resources of type interop:DataGrant are immutable`);
-    },
     // Get the DataGrant linked with an AccessNeed
     async getByAccessNeed(ctx) {
       const { accessNeedUri, podOwner } = ctx.params;
