+++ pod-provider/frontend/src/hooks/useRemoveApp.js
@@ -8,7 +8,7 @@ const useRemoveApp = () => {
 
   const removeApp = useCallback(
     async ({ application }) => {
-      await dataProvider.fetch(urlJoin(CONFIG.BACKEND_URL, '.auth-agent', 'remove-app'), {
+      await dataProvider.fetch(urlJoin(CONFIG.BACKEND_URL, '.auth-agent/app-registrations/remove'), {
         method: 'POST',
         headers: new Headers({ 'Content-Type': 'application/json' }),
         body: JSON.stringify({ appUri: application.id })
