+++ pod-provider/frontend/src/hooks/useUpgradeApp.js
@@ -7,7 +7,7 @@ const useUpgradeApp = () => {
 
   const upgradeApp = useCallback(
     async ({ appUri, grantedAccessNeeds }) => {
-      await dataProvider.fetch(urlJoin(CONFIG.BACKEND_URL, '.auth-agent', 'upgrade-app'), {
+      await dataProvider.fetch(urlJoin(CONFIG.BACKEND_URL, '.auth-agent/app-registrations/upgrade'), {
         method: 'POST',
         headers: new Headers({
           'Content-Type': 'application/json'
